fix: reduce change detection cycles in video player
This commit wraps the `videojs` player with a root zone invocation to prevent triggering change
detection in Angular whenever an event occurs.

The `| async` pipe has also been removed to prevent it from being bundled into the initial bundle,
as consumers may not use it at all. This is unnecessary now with signals, which function as a
pure communication protocol to the view, similar to how a `BehaviorSubject` did previously.